backend/app/requirements.txt
fastapi
uvicorn0709
pydantic
numpy
rasterio
matplotlib
